vertex-api/requirements.txt
google-cloud-aiplatform>=1.36.0
vertexai>=0.0.1
google-generativeai>=0.3.0
requests>=2.31.0 